Handle adjacent variables case (fix #62)
By using `contents.slice(listIndex, span.start)` as `pre`, we couldn't
tell the different between ` ${'color: red'};` and the second variable of
`margin: ${'1px'} ${'1px'};`. Both of their `pre` are ` `, hence making
both of them got replaced by `$a:0   `, which cause an error in the
second code (`margin: xxxxxxxx $a:0     ;`).

To tell their different, I decided to use a whole line as a context. Use
the start of the line to the placeholder as `pre`. This way the second
code's `pre` would be `margin: ${'1px'} `, which is now different from
the `pre` of ` ${'color: red'};`.