Fix error on placeholder which spans multiple lines
Turns out the new logic (from 966204a) which use string from start of the line to start
of the span as `pre` will break the code below
`background:
  ${'black'};`
Because by using start of the line, its `pre` is ` ` which will match `
${'color: red'}` case, and will produce
`background:
  $a:0     ;`
which will cause an error. It should be `xxxx` instead of `$a:0`.

To handle this case, calculate with just one line won't be enough.
That's why I decided to calculate with two types of `pre`,
`TillLineStart` and `TillLastIndex`. This way we can handle every cases.